Meshes extension: Fixed transparent backfaces, tidy up
